docs: clarify node sdk instructions wrt envs and tokens
## What

This change does two things:

It **removes the `environment` option** from the initialization
examples. This option is deprecated and only causes confusion. There's
no good reason to have it lying around in the examples.

It also **updates the explanation of the code samples**, telling you
that you need to generate an API key and linking you to the reference
doc and how to guide.

Relates to: #2782.