Fix(typo): Fix "login"/"log in", "logout"/"log out" issues
## What

This change updates some places in the docs where we use the terms
"login" and "logout" incorrectly.

A "login" is a noun, typically referring to the set of credentials you
need to _log in_ to a service. The verb form, the act of signing in,
is written in two words: to "log in".

A similar logic applies to "logout" and "log out", although I don't
find the term "logout" in my dictionary. However, I think it makes
sense to talk about "logout requests" (and I see references to logout
in other services and documentation), so I'm happy to use that as a
noun. Regardless, the act of logging out is to "log out".