docs: fix links that broke due to the doc move
Because docs now live in a different folder, links to the
documents (that have been moved) must be updated.

I've chosen the use the markdown link for the generated docs for now
instead of the URL because docusaurus will alert us when something
breaks. However, if that proves to be a pain, we can always use the
url instead.