docs: Add front-end API setup to the quickstart guide
## What

This change adds instructions on how to use the front-end API for
client-side SDKs to the quickstart guide. It also organizes the
content so that this info is presented before the proxy info.

## Why

The front-end API is easier to use, so it should be preferred over the
proxy for people who are just getting started.