docs: add notes to constraints and segment docs about large values (#3142)
## What

This PR adds a note to the constraint and the segment docs about using a
large number of constraint values.

## Why

This is a fairly common question in slack, and there is no clear
documentation around this at the moment. As such, it's likely that
people do this a fair bit without telling us. We don't think this is the
right approach, so adding this bit of documentation should hopefully
make it a little clearer.