docs: clarify that we don't support multi-host db urls (#3312)
This PR updates the docs to clarify that we don't support multi-host
strings in the databaseUrl configuration variable.

The docs used to just mention that we accept libpq strings, but as
reported in #3221, we **don't** support multi-host strings.

I did some digging and found that this is because [node-postgres -- the
library we use to connect to postgres -- itself does not support
multi-host
strings](https://github.com/brianc/node-postgres/issues/2308).

To fix this, we would need to replace the library we use or otherwise
wrap it. While this would be a better long-term solution, we don't have
the capacity to it right now. But we can update the docs to better
reflect reality.