chore: remove uses of type errors from user-facing code
This change removes uses of the `TypeError` type from user-facing
code. Type errors are used by typescript when you provide it the wrong
type. This is a valid concern. However, in the API, they're usually a
signal that we've done something wrong rather than the user. As such,
it makes more sense to return them as validation errors or bad request errors.